Fix int-to-pointer-cast warnings for MIPS.

+++ ports/sysdeps/unix/sysv/linux/mips/____longjmp_chk.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 2009 Free Software Foundation, Inc.
+/* Copyright (C) 2009-2012 Free Software Foundation, Inc.
    This file is part of the GNU C Library.
 
    The GNU C Library is free software; you can redistribute it and/or
@@ -29,7 +29,7 @@
 	int ret = __sigaltstack (NULL, &ss);				\
 	if (ret == 0							\
 	    && (!(ss.ss_flags & SS_ONSTACK)				\
-		|| ((unsigned sp_type) ((sp_type) ss.ss_sp		\
+		|| ((unsigned sp_type) ((sp_type) (long) ss.ss_sp	\
 					+ (sp_type) ss.ss_size		\
 					- sp_saved)			\
 		    < ss.ss_size)))					\
